ライトの共通化
