+++ project/Engine/3D/Object3D.cpp
@@ -18,31 +18,6 @@ Object3D::Object3D()
 	materialCB_.data_->uvTransform = Matrix::Identity();
 	// 光沢を初期化
 	materialCB_.data_->shininess = 50.0f;
-
-	// 平行光源のデフォルト値を書き込む
-	directionalLightCB_.data_->color = { 1.0f,1.0f,1.0f,1.0f };
-	directionalLightCB_.data_->direction = { 0.0f, -1.0f, 0.0f };
-	directionalLightCB_.data_->intensity = 0.1f;   // 平行光源有効
-	/*directionalLightCB_.data_->intensity = 0.0f;*/   // 平行光源無効
-
-	// ポイントライトのデフォルト値を書き込む
-	pointLightCB_.data_->color = { 1.0f, 1.0f, 1.0f, 1.0f };
-	pointLightCB_.data_->position = { 0.0f, 2.0f, 0.0f };
-	/*pointLightCB_.data_->intensity = 1.0f; */   // ポイントライト有効
-	pointLightCB_.data_->intensity = 0.0f;    // ポイントライト無効
-	pointLightCB_.data_->radius = 5.0f;
-	pointLightCB_.data_->decay = 1.0f;
-
-	// スポットライトのデフォルト値を書き込む
-	spotLightCB_.data_->color = {1.0f, 1.0f, 1.0f, 1.0f};
-	spotLightCB_.data_->position = {2.0f, 1.25f, 0.0f};
-	spotLightCB_.data_->distance = 7.0f;
-	spotLightCB_.data_->direction = {-1.0f, -1.0f, 0.0f};
-	spotLightCB_.data_->intensity = 4.0f; // スポットライト有効
-	/*spotLightCB_.data_->intensity = 0.0f;*/ // スポットライト無効
-	spotLightCB_.data_->decay = 2.0f;
-	spotLightCB_.data_->cosAngle = std::cos(std::numbers::pi_v<float> / 3.0f);
-	spotLightCB_.data_->cosFalloffStart = std::cos(std::numbers::pi_v<float> / 4.0f);
 }
 
 void Object3D::UpdateMatrix()
@@ -65,13 +40,6 @@ void Object3D::Draw()
 {
 	DirectXBase* dxBase = DirectXBase::GetInstance();
 
-	// 平行光源の定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(3, directionalLightCB_.resource_->GetGPUVirtualAddress());
-	// ポイントライトの定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(6, pointLightCB_.resource_->GetGPUVirtualAddress());
-	// スポットライトの定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(7, spotLightCB_.resource_->GetGPUVirtualAddress());
-
 	// commandListにVBVを設定
 	dxBase->GetCommandList()->IASetVertexBuffers(0, 1, &model_->vertexBufferView);
 	// commandListにIBVを設定
@@ -92,13 +60,6 @@ void Object3D::Draw(const int TextureHandle)
 {
 	DirectXBase* dxBase = DirectXBase::GetInstance();
 
-	// 平行光源の定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(3, directionalLightCB_.resource_->GetGPUVirtualAddress());
-	// ポイントライトの定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(6, pointLightCB_.resource_->GetGPUVirtualAddress());
-	// スポットライトの定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(7, spotLightCB_.resource_->GetGPUVirtualAddress());
-
 	// commandListにVBVを設定
 	dxBase->GetCommandList()->IASetVertexBuffers(0, 1, &model_->vertexBufferView);
 	// マテリアルCBufferの場所を設定
@@ -114,13 +75,6 @@ void Object3D::Draw(const int TextureHandle)
 void Object3D::Draw(ModelManager::SkinCluster skinCluster)
 {
 	DirectXBase* dxBase = DirectXBase::GetInstance();
-
-	// 平行光源の定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(3, directionalLightCB_.resource_->GetGPUVirtualAddress());
-	// ポイントライトの定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(6, pointLightCB_.resource_->GetGPUVirtualAddress());
-	// スポットライトの定数バッファをセット
-	dxBase->GetCommandList()->SetGraphicsRootConstantBufferView(7, spotLightCB_.resource_->GetGPUVirtualAddress());
 
 	D3D12_VERTEX_BUFFER_VIEW vbvs[2] = {
 		model_->vertexBufferView, // VertexDataのVBV
