+++ project/Engine/3D/Object3D.h
@@ -22,33 +22,6 @@ public:
 		Matrix WorldInverseTranspose;
 	};
 
-	struct DirectionalLight {
-		Float4 color; // ライトの色
-		Float3 direction; // ライトの向き
-		float intensity; // 輝度
-	};
-
-	struct PointLight {
-		Float4 color; // ライトの色
-		Float3 position; // ライトの位置
-		float intensity; // 輝度
-		float radius; // ライトの届く最大距離
-		float decay; // 減衰率
-		float padding[2];
-	};
-
-	struct SpotLight {
-		Float4 color;     // ライトの色
-		Float3 position;  // ライトの位置
-		float intensity;  // 輝度
-		Float3 direction; // スポットライトの方向
-		float distance;   // ライトの届く最大距離
-		float decay;      // 減衰率
-		float cosAngle;   // スポットライトの余弦
-		float cosFalloffStart; // Falloff開始の角度
-		float padding[1];
-	};
-
 	struct ParticleForGPU {
 		Matrix WVP;
 		Matrix World;
@@ -80,13 +53,5 @@ public:
 
 	// トランスフォーム情報
 	Transform transform_;
-
-	// 平行光源の定数バッファ
-	ConstBuffer<DirectionalLight> directionalLightCB_;
-	// ポイントライトの定数バッファ
-	ConstBuffer<PointLight> pointLightCB_;
-	// スポットライトの定数バッファ
-	ConstBuffer<SpotLight> spotLightCB_;
-
 };
