+++ project/externals/DirectXTex/Bin/Desktop_2022_Win10/x64/Debug/DirectXTex.tlog/Cl.items.tlog
@@ -1,21 +1,21 @@
-C:\school\Engine\project\externals\DirectXTex\BC.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BC.obj
-C:\school\Engine\project\externals\DirectXTex\BC4BC5.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BC4BC5.obj
-C:\school\Engine\project\externals\DirectXTex\BC6HBC7.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BC6HBC7.obj
-C:\school\Engine\project\externals\DirectXTex\BCDirectCompute.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BCDirectCompute.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexCompress.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexCompress.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexCompressGPU.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexCompressGPU.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexConvert.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexConvert.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexD3D11.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexD3D11.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexD3D12.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexD3D12.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexDDS.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexDDS.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexFlipRotate.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexFlipRotate.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexHDR.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexHDR.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexImage.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexImage.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexMipMaps.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexMipMaps.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexMisc.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexMisc.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexNormalMaps.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexNormalMaps.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexPMAlpha.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexPMAlpha.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexResize.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexResize.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexTGA.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexTGA.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexUtil.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexUtil.obj
-C:\school\Engine\project\externals\DirectXTex\DirectXTexWIC.cpp;C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexWIC.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\BC.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BC.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\BC4BC5.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BC4BC5.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\BC6HBC7.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BC6HBC7.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\BCDirectCompute.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\BCDirectCompute.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexCompress.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexCompress.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexCompressGPU.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexCompressGPU.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexConvert.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexConvert.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexD3D11.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexD3D11.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexD3D12.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexD3D12.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexDDS.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexDDS.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexFlipRotate.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexFlipRotate.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexHDR.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexHDR.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexImage.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexImage.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexMipMaps.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexMipMaps.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexMisc.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexMisc.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexNormalMaps.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexNormalMaps.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexPMAlpha.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexPMAlpha.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexResize.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexResize.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexTGA.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexTGA.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexUtil.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexUtil.obj
+D:\k023g0121\Engine\project\externals\DirectXTex\DirectXTexWIC.cpp;D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTexWIC.obj
