+++ project/externals/DirectXTex/Bin/Desktop_2022_Win10/x64/Debug/DirectXTex.tlog/DirectXTex.lastbuildstate
@@ -1,2 +1,2 @@
 PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.39.33519:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\school\Engine\project\|
+Debug|x64|D:\k023g0121\Engine\project\|
