+++ project/externals/DirectXTex/Bin/Desktop_2022_Win10/x64/Debug/DirectXTex_Desktop_2022_Win10.log
@@ -20,4 +20,4 @@
   DirectXTexTGA.cpp
   DirectXTexWIC.cpp
   コードを生成中...
-  DirectXTex_Desktop_2022_Win10.vcxproj -> C:\school\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTex.lib
+  DirectXTex_Desktop_2022_Win10.vcxproj -> D:\k023g0121\Engine\project\externals\DirectXTex\Bin\Desktop_2022_Win10\x64\Debug\DirectXTex.lib
